Pr: Probe with Retries (#756)
* Stop probe moving the toolhead in x/y when it retracts

Implement a single `_retract` method so it is always performed consistently. This prepares for probe retry strategies that move the probe in x/y.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

* Give 'probing_move' implementions access to the GCode Command object

Pass the GCodeCommand object to the `probing_move` method. This allows probes access to any probe specific custom parameter that might be in the command when probing.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

* Cleanup drop first result behaviour

This has diplicated code and is unecessarily complex leading to further complexity in later changes.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

* Implement probing retries for fouled probes

Nozzle probes suffer from ooze. If a probe can detect that the nozzle was fouled it can report that infromation back to the probing system. This change adds an optional return parameter, `is_good`, to the `probing_move` interface. If a probe is not good, it can take action to execute additional probes to resolve the issue.

This is all under control of the user via the retry strategy. The strategies are: FAIL, IGNORE, RETRY and CIRCLE. RETRY and CIRCLE re-attempt probes with CIRCLE moving the probe to a clean location in a very small circle around the original probe location.

CIRCLE is the preferred strategy for bed meshing. FAIL or RETRY might be a good strategy for something like QGL where the absolute position of the probe is critical. IGNORE is a good strategy when probing things that are not as rigid as the bed, such as a nozzle scrubber.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

* Use a RetrySession in ProbePointsHelper

If using the CIRCLE stratey and retrying multiple times at the same points (e.g. for QGL) any fouled points are rememberd and avoided by keeping a session between attempts (calls to `start_probe`). The number of retries are reset for all points on each pass but the fouled points are saved.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

* Implement nozzle_scrubber_gcode and SCRUBBING_FREQUENCY options.

Allow the user to configure their own custom nozzle scrubbing routine that can integrate with custom nozzle scrubbing hardware, such as a brush or wiper. This is invoked when a probe fails. The SCRUBBING_FREQUENCY parameter allows this to be combined with retry strategies, particularly the CIRCLE and RETRY strategies. This can provide enhanced intermittent scrubbing if there is a problem with tapping to clear a fouled nozzle.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

* Home from last probe position

This is a bit of a tricky/archane macro to write so this saves you the trouble. PROBE HOME=z perfectly homes your Z axis if you have a nozzle probe.

HOME=z was used so later we can probe in x/y and submit alternate axis names.


Signed-off-by: Gareth Farrington <gareth@waves.ky>

---------

Signed-off-by: Gareth Farrington <gareth@waves.ky>